Add login and rename routes
