+++ templates/Back/index.html.twig
@@ -1,4 +1,4 @@
-{% extends 'back/layouts/app.html.twig' %}
+{% extends 'Back/layouts/app.html.twig' %}
 
 {% block body %}
     My back Body
