+++ templates/Front/index.html.twig
@@ -1,4 +1,4 @@
-{% extends 'back/layouts/app.html.twig' %}
+{% extends 'Front/layouts/app.html.twig' %}
 
 {% block body %}
     My front Body
