+++ templates/Front/layouts/app.html.twig
@@ -6,11 +6,9 @@
         {% block stylesheets %}{% endblock %}
     </head>
     <body>
-        {% include "back/layouts/navbar.html.twig" %}
-
+        {% include "Front/layouts/navbar.html.twig" %}
         {% block breadcrum %}{% endblock %}
         {% block body %}{% endblock %}
-
         {% block javascripts %}{% endblock %}
     </body>
 </html>
