+++ templates/security/login.html.twig
@@ -1,4 +1,4 @@
-{% extends 'base.html.twig' %}
+{% extends 'Front/index.html.twig' %}
 
 {% block title %}Log in!{% endblock %}
 
@@ -10,7 +10,7 @@
 
     <h1 class="h3 mb-3 font-weight-normal">Please sign in</h1>
     <label for="inputEmail" class="sr-only">Email</label>
-    <input type="email" value="{{ last_username }}" name="email" id="inputEmail" class="form-control" placeholder="Email" required autofocus>
+    <input type="email" name="email" id="inputEmail" class="form-control" placeholder="Email" required autofocus>
     <label for="inputPassword" class="sr-only">Password</label>
     <input type="password" name="password" id="inputPassword" class="form-control" placeholder="Password" required>
